Move type to store
[MAILPOET-2911]
